LOGBROKER-10206 Fix
